Cleanup
Added propper repository and DataAccessObject Interfaces
