--- EK.Discord.Server.Notion/Base/Persistence/IEntity.cs
@@ -1,13 +1,0 @@
-﻿namespace EK.Discord.Server.Notion.Base.Persistence;
-
-public interface IEntity {
-
-}
-
-public interface IEntity<out TKey> : IEntity {
-    public TKey Id { get; }
-}
-
-public interface IGuidEntity : IEntity<Guid> {
-    
-}
