+++ EK.Discord.Server/Program.cs
@@ -1,11 +1,8 @@
 using System.Diagnostics.CodeAnalysis;
-using Discord;
 using EK.Discord.Server.Bot.Base.DependencyInjection;
 using EK.Discord.Server.Base.Configuration;
-using EK.Discord.Server.Base.Notion;
-using EK.Discord.Server.Notion;
+using EK.Discord.Server.Notion.Base;
 using EK.Discord.Server.TemplateComponent;
-using Notion.Client;
 
 namespace EK.Discord.Server;
